chore: move PolicyManager into policy pkg
