+++ pkg/policy/manager_test.go
@@ -1,4 +1,4 @@
-package ebpf
+package policy
 
 import (
 	"sync"
@@ -7,13 +7,12 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/aquasecurity/tracee/pkg/events"
-	"github.com/aquasecurity/tracee/pkg/policy"
 )
 
 func TestPolicyManagerEnableRule(t *testing.T) {
 	t.Parallel()
 
-	policyManager := newPolicyManager()
+	policyManager := NewPolicyManager()
 
 	policy1Mached := uint64(0b10)
 	policy2Mached := uint64(0b100)
@@ -39,7 +38,7 @@ func TestPolicyManagerEnableRule(t *testing.T) {
 func TestPolicyManagerDisableRule(t *testing.T) {
 	t.Parallel()
 
-	policyManager := newPolicyManager()
+	policyManager := NewPolicyManager()
 
 	policy1Mached := uint64(0b10)
 	policy2Mached := uint64(0b100)
@@ -77,13 +76,13 @@ func TestPolicyManagerEnableAndDisableRuleConcurrent(t *testing.T) {
 		events.FileModification,
 	}
 
-	policyManager := newPolicyManager()
+	policyManager := NewPolicyManager()
 
 	var wg sync.WaitGroup
 
 	wg.Add(1)
 	go func() {
-		for i := 0; i < policy.PolicyMax; i++ {
+		for i := 0; i < PolicyMax; i++ {
 			for _, e := range eventsToEnable {
 				policyManager.EnableRule(i, e)
 			}
@@ -93,7 +92,7 @@ func TestPolicyManagerEnableAndDisableRuleConcurrent(t *testing.T) {
 
 	wg.Add(1)
 	go func() {
-		for i := 0; i < policy.PolicyMax; i++ {
+		for i := 0; i < PolicyMax; i++ {
 			for _, e := range eventsToDisable {
 				policyManager.DisableRule(i, e)
 			}
@@ -103,12 +102,12 @@ func TestPolicyManagerEnableAndDisableRuleConcurrent(t *testing.T) {
 
 	wg.Wait()
 
-	for i := 0; i < policy.PolicyMax; i++ {
+	for i := 0; i < PolicyMax; i++ {
 		for _, e := range eventsToEnable {
-			assert.True(t, policyManager.IsRuleEnabled(policy.PolicyAll, e))
+			assert.True(t, policyManager.IsRuleEnabled(PolicyAll, e))
 		}
 		for _, e := range eventsToDisable {
-			assert.False(t, policyManager.IsRuleEnabled(policy.PolicyAll, e))
+			assert.False(t, policyManager.IsRuleEnabled(PolicyAll, e))
 		}
 	}
 }
@@ -116,7 +115,7 @@ func TestPolicyManagerEnableAndDisableRuleConcurrent(t *testing.T) {
 func TestPolicyManagerEnableEvent(t *testing.T) {
 	t.Parallel()
 
-	policyManager := newPolicyManager()
+	policyManager := NewPolicyManager()
 
 	assert.False(t, policyManager.isEventEnabled(events.SecurityBPF))
 	assert.False(t, policyManager.isEventEnabled(events.SecurityFileOpen))
@@ -134,7 +133,7 @@ func TestPolicyManagerEnableEvent(t *testing.T) {
 func TestPolicyManagerDisableEvent(t *testing.T) {
 	t.Parallel()
 
-	policyManager := newPolicyManager()
+	policyManager := NewPolicyManager()
 
 	policyManager.EnableEvent(events.SecurityBPF)
 	policyManager.EnableEvent(events.SecurityFileOpen)
@@ -171,7 +170,7 @@ func TestPolicyManagerEnableAndDisableEventConcurrent(t *testing.T) {
 		events.FileModification,
 	}
 
-	policyManager := newPolicyManager()
+	policyManager := NewPolicyManager()
 
 	// activate events
 	for _, e := range eventsToDisable {
@@ -182,7 +181,7 @@ func TestPolicyManagerEnableAndDisableEventConcurrent(t *testing.T) {
 
 	wg.Add(1)
 	go func() {
-		for i := 0; i < policy.PolicyMax; i++ {
+		for i := 0; i < PolicyMax; i++ {
 			for _, e := range eventsToEnable {
 				policyManager.EnableEvent(e)
 			}
@@ -192,7 +191,7 @@ func TestPolicyManagerEnableAndDisableEventConcurrent(t *testing.T) {
 
 	wg.Add(1)
 	go func() {
-		for i := 0; i < policy.PolicyMax; i++ {
+		for i := 0; i < PolicyMax; i++ {
 			for _, e := range eventsToDisable {
 				policyManager.DisableEvent(e)
 			}
@@ -202,7 +201,7 @@ func TestPolicyManagerEnableAndDisableEventConcurrent(t *testing.T) {
 
 	wg.Wait()
 
-	for i := 0; i < policy.PolicyMax; i++ {
+	for i := 0; i < PolicyMax; i++ {
 		for _, e := range eventsToEnable {
 			assert.True(t, policyManager.IsEventEnabled(e))
 		}
@@ -215,7 +214,7 @@ func TestPolicyManagerEnableAndDisableEventConcurrent(t *testing.T) {
 func TestEnableRuleAlsoEnableEvent(t *testing.T) {
 	t.Parallel()
 
-	policyManager := newPolicyManager()
+	policyManager := NewPolicyManager()
 
 	assert.False(t, policyManager.IsEventEnabled(events.SecurityBPF))
 
@@ -227,7 +226,7 @@ func TestEnableRuleAlsoEnableEvent(t *testing.T) {
 func TestDisableRuleAlsoEnableEvent(t *testing.T) {
 	t.Parallel()
 
-	policyManager := newPolicyManager()
+	policyManager := NewPolicyManager()
 
 	assert.False(t, policyManager.IsEventEnabled(events.SecurityFileOpen))
 
@@ -239,7 +238,7 @@ func TestDisableRuleAlsoEnableEvent(t *testing.T) {
 func TestPolicyManagerIsEnabled(t *testing.T) {
 	t.Parallel()
 
-	policyManager := newPolicyManager()
+	policyManager := NewPolicyManager()
 
 	policy1Mached := uint64(0b10)
 	policy2Mached := uint64(0b100)
